perf(SearchBar): Switch from material-community to material (#1620)
Changes the SearchBar icons from material-community to material.

The default icon is material. We should use it here too, so RNW does not have to load two fonts just because the SearchBar uses another one than the default.

This increases the performance of the app and needs less network traffic because only one font is needed by default.